Make the stk500 protocol commands stateless
Move the state out of the protocol writes. This makes the logic clearer
but messes up the error handling (before errors would return to the
caller; now the caller will get a mysterious exception).

Argument handling for the cmd/2 functions uses a record. Protocol
commands will be exported to the library user but without forcing the
user to pass in a record.